clean parse_*.c
