+++ srcs/ogv1.c
@@ -90,7 +90,7 @@ void	initialize_render(t_mlx *mlx, t_loader loader)
 	mlx->window = mlx_new_window(mlx->handle, WINDOW_WIDTH, WINDOW_HEIGHT, "cuv3d"); 
 	mlx->img.handle = mlx_new_image(mlx->handle, WINDOW_WIDTH, WINDOW_HEIGHT);
 	mlx->img.buffer = mlx_get_data_addr(mlx->img.handle, &(mlx->img.bits_per_pixel), &(mlx->img.line_size), &(mlx->img.endian));
-	mlx->player.pos = convert_player_pos(loader);
+	mlx->player.pos = loader.player_pos;//convert_player_pos(loader)
 	mlx->player.angle = loader.player_angle;
 	mlx->map.data = ft_double_str_dup(loader.map_data);
 	mlx->map.y_count = loader.map_y_count;
